bug fixing-03

+++ frontend/src/app/layout/charts/main-catagory/main-catagory.component.ts
@@ -5,7 +5,7 @@ import { AuthenticationService } from 'src/app/services/authentication.service';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { MainNewCatagoryComponent } from './main-new-catagory/main-new-catagory.component';
 import { MessageAlertDisplay } from 'src/app/common-class/message-alert-display';
-
+import Swal from 'sweetalert2/dist/sweetalert2.js';
 @Component({
   selector: 'app-main-catagory',
   templateUrl: './main-catagory.component.html',
@@ -41,21 +41,39 @@ export class MainCatagoryComponent implements OnInit {
   }
 
   deleteRecord(data) {
-    console.log(data._id);
-    let idData = {
-      "id" : data._id
-    }
-    this.catagoryService.deleteRecord(idData)
-    .subscribe(data => {
-      console.log(data);
-      refresh:true;
-    },
-    error => {
-      this.error = error;
-      this.loading = false;
 
-    });
-    location.reload();
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this imaginary file!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+      Swal.fire(
+        'Deleted!',
+        'Your imaginary file has been deleted.',
+        'success'
+      );
+      let idData = {
+        "id" : data._id
+      }
+      this.catagoryService.deleteRecord(idData)
+      .subscribe(data => {
+        console.log(data);
+        refresh:true;
+      },
+      error => {
+        this.error = error;
+        this.loading = false;
+  
+      });
+      this.loadData();
+    } })
+
+
   }
 
   openModal() {
